requirements.txt
google-generativeai
qdrant-client
streamlit
fastembed
dotenv